ДЗ.3. Реализовано. Исправлены стилистические недочеты

+++ admin/delete.php
@@ -4,6 +4,8 @@ include __DIR__ . '/../App/autoload.php';
 
 if (isset($_GET['id'])){
     $article = \App\Models\Article::findById($_GET['id']);
-    $article->delete();
+    if ($article){
+        $article->delete();
+    }
     header('Location: /admin/');
 }
